Add firmware declarations for ipts_surface
Declaring the required firmware files as MODULE_FIRMWARE makes it possible to
include the driver in the initramfs, in case that is needed or desired.

Even though there are files that are loaded through a config file, that config
file is the same for every Surface using IPTS, so this should be fine.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>